+++ lib/alerts.py
@@ -1,11 +1,11 @@
 """
 /lib/alerts.py
-Server Monitoring System v8.33.31
+Server Monitoring System v8.33.32
 Copyright (c) 2025 Aleksandr Sukhanov
 License: MIT
 Unified alert system
 Система мониторинга серверов
-Версия: 8.33.31
+Версия: 8.33.32
 Автор: Александр Суханов (c)
 Лицензия: MIT
 Единая система оповещений