+++ modules/resources.py
@@ -1,11 +1,11 @@
 """
 /app/modules/resources.py
-Server Monitoring System v8.33.31
+Server Monitoring System v8.33.32
 Copyright (c) 2025 Aleksandr Sukhanov
 License: MIT
 Server resource checking module
 Система мониторинга серверов
-Версия: 8.33.31
+Версия: 8.33.32
 Автор: Александр Суханов (c)
 Лицензия: MIT
 Модуль проверки ресурсов серверов